Complete 5/6 stage, project Web Engine Quiz

+++ Web Quiz Engine/Web Quiz Engine/task/src/engine/businesslayer/services/QuizService.java
@@ -5,6 +5,9 @@ import engine.persistence.QuizRepository;
 import engine.presentation.Answer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -34,6 +37,8 @@ public class QuizService {
     }
 
     public void saveNewQuizzes(Quiz quiz) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        quiz.setCreator(authentication.getName());
         quizRepository.save(quiz);
     }
 
@@ -52,10 +57,13 @@ public class QuizService {
                 new Answer(false, "Wrong answer! Please, try again.");
     }
 
-    public ResponseStatusException delete(Long id) {
+    public ResponseEntity delete(Long id) {
         if (quizRepository.findById(id).isPresent()) {
-            quizRepository.deleteById(id);
-            return new ResponseStatusException(HttpStatus.NO_CONTENT);
-        } else return new ResponseStatusException(HttpStatus.NOT_FOUND);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (quizRepository.findById(id).get().getCreator().equals(authentication.getName())) {
+                quizRepository.deleteById(id);
+            } else return new ResponseEntity(HttpStatus.FORBIDDEN);
+            return new ResponseEntity(HttpStatus.NO_CONTENT);
+        } else return new ResponseEntity(HttpStatus.NOT_FOUND);
     }
 }
